+++ src/index.js
@@ -3,9 +3,10 @@ import Rendering from './rendering';
 import IO from './io';
 import Element from './element';
 import Filters from './filters';
+import Image from './image';
 import constants from './constants';
 
-(function(p5) {
+function init(p5) {
     /**
      * @namespace p5
      */
@@ -14,9 +15,16 @@ import constants from './constants';
     IO(p5);
     Element(p5);
     Filters(p5);
+    Image(p5);
 
     // attach constants to p5 instance
     Object.keys(constants).forEach(function(k) {
         p5.prototype[k] = constants[k];
     });
-})(window.p5);
+}
+
+if (typeof window.p5 !== 'undefined') {
+    init(window.p5);
+}
+
+export default init;