+++ src/p5.RendererSVG.js
@@ -1,6 +1,7 @@
-import {Element as SVGCanvasElement} from 'svgcanvas';
+import { Element as SVGCanvasElement } from 'svgcanvas';
+import { DEBUG } from './config';
 
-export default function(p5) {
+export default function (p5) {
     /**
      * @namespace RendererSVG
      * @constructor
@@ -9,7 +10,7 @@ export default function(p5) {
      * @param {Bool} isMainCanvas
      */
     function RendererSVG(elt, pInst, isMainCanvas) {
-        var svgCanvas = new SVGCanvasElement();
+        var svgCanvas = new SVGCanvasElement({ debug: DEBUG });
         var svg = svgCanvas.svg;
 
         // replace <canvas> with <svg> and copy id, className
@@ -23,7 +24,7 @@ export default function(p5) {
 
         elt.parentNode = {
             // fake parentNode.removeChild so that noCanvas will work
-            removeChild: function(element) {
+            removeChild: function (element) {
                 if (element === elt) {
                     var wrapper = svgCanvas.getElement();
                     wrapper.parentNode.removeChild(wrapper);
@@ -31,37 +32,37 @@ export default function(p5) {
             }
         };
 
-        p5.Renderer2D.call(this, elt, pInst, isMainCanvas);
+        const pInstProxy = new Proxy(pInst, {
+            get: function (target, prop) {
+                if (prop === '_pixelDensity') {
+                    // 1 is OK for SVG
+                    return 1;
+                }
+                return target[prop];
+            }
+        });
+
+        p5.Renderer2D.call(this, elt, pInstProxy, isMainCanvas);
 
         this.isSVG = true;
         this.svg = svg;
+
+        if (DEBUG) {
+            console.debug({ svgCanvas });
+            console.debug(this.drawingContext);
+        }
 
         return this;
     }
 
     RendererSVG.prototype = Object.create(p5.Renderer2D.prototype);
 
-    RendererSVG.prototype._applyDefaults = function() {
+    RendererSVG.prototype._applyDefaults = function () {
         p5.Renderer2D.prototype._applyDefaults.call(this);
         this.drawingContext.lineWidth = 1;
     };
 
-    RendererSVG.prototype.line = function(x1, y1, x2, y2) {
-        var styleEmpty = 'rgba(0,0,0,0)';
-        var ctx = this.drawingContext;
-        if (!this._doStroke) {
-            return this;
-        } else if(ctx.strokeStyle === styleEmpty){
-            return this;
-        }
-        ctx.beginPath();
-        ctx.moveTo(x1, y1);
-        ctx.lineTo(x2, y2);
-        ctx.stroke();
-        return this;
-    };
-
-    RendererSVG.prototype.resize = function(w, h) {
+    RendererSVG.prototype.resize = function (w, h) {
         if (!w || !h) {
             return;
         }
@@ -71,35 +72,17 @@ export default function(p5) {
             // note that at first this.width and this.height is undefined
             this.drawingContext.__clearCanvas();
         }
-        this._withPixelDensity(function() {
-            p5.Renderer2D.prototype.resize.call(this, w, h);
-        });
+
+        p5.Renderer2D.prototype.resize.call(this, w, h);
+
         // For scale, crop
         // see also: http://sarasoueidan.com/blog/svg-coordinate-systems/
         this.svg.setAttribute('viewBox', [0, 0, w, h].join(' '));
     };
 
-    /**
-     * @private
-     */
-    RendererSVG.prototype._withPixelDensity = function(fn) {
-        let pixelDensity = this._pInst._pixelDensity;
-        this._pInst._pixelDensity = 1; // 1 is OK for SVG
-        fn.apply(this);
-        this._pInst._pixelDensity = pixelDensity;
-    };
-
-    RendererSVG.prototype.background = function() {
-        var args = arguments;
-        this._withPixelDensity(function() {
-            p5.Renderer2D.prototype.background.apply(this, args);
-        });
-    };
-
-    RendererSVG.prototype.resetMatrix = function() {
-        this._withPixelDensity(function() {
-            p5.Renderer2D.prototype.resetMatrix.apply(this);
-        });
+    RendererSVG.prototype.clear = function () {
+        p5.Renderer2D.prototype.clear.call(this);
+        this.drawingContext.__clearCanvas();
     };
 
     /**
@@ -109,7 +92,7 @@ export default function(p5) {
      * @memberof RendererSVG.prototype
      * @param {SVGElement|Element} element
      */
-    RendererSVG.prototype.appendChild = function(element) {
+    RendererSVG.prototype.appendChild = function (element) {
         if (element && element.elt) {
             element = element.elt;
         }
@@ -130,7 +113,7 @@ export default function(p5) {
      * @param {Number} width
      * @param {Number} height
      */
-    RendererSVG.prototype.image = function(img, sx, sy, sWidth, sHeight, x, y, w, h) {
+    RendererSVG.prototype.image = function (img, sx, sy, sWidth, sHeight, x, y, w, h) {
         if (!img) {
             throw new Error('Invalid image: ' + img);
         }
@@ -149,11 +132,39 @@ export default function(p5) {
                 sHeight /= this._pInst._pixelDensity;
                 elt.setAttribute('viewBox', [sx, sy, sWidth, sHeight].join(', '));
             }
-            this.appendChild(elt);
+
+            let g = p5.SVGElement.create('g');
+            this.drawingContext.__applyTransformation(g.elt);
+            g.elt.appendChild(elt);
+            this.appendChild(g.elt);
         } else {
             p5.Renderer2D.prototype.image.apply(this, arguments);
         }
     };
+
+    /**
+     * @method parent
+     * @return {p5.Element}
+     *
+     * @see https://github.com/zenozeng/p5.js-svg/issues/187
+     */
+    RendererSVG.prototype.parent = function () {
+        const $this = {
+            elt: this.elt.getElement()
+        };
+        return p5.Element.prototype.parent.apply($this, arguments);
+    };
+
+
+    RendererSVG.prototype.loadPixels = async function () {
+        const pixelsState = this._pixelsState; // if called by p5.Image
+        const pd = pixelsState._pixelDensity;
+        const w = this.width * pd;
+        const h = this.height * pd;
+        const imageData = await this.drawingContext.getImageData(0, 0, w, h, { async: true });
+        pixelsState._setProperty('imageData', imageData);
+        pixelsState._setProperty('pixels', imageData.data);
+    }
 
     p5.RendererSVG = RendererSVG;
 }